+++ src/oracle.rs
@@ -1,7 +1,7 @@
+use anyhow::Result;
 use structopt::StructOpt;
 use terra_rust_api::Terra;
-
-use crate::errors::Result;
+//use crate::errors::Result;
 //use crate::keys::get_private_key;
 use crate::{NAME, VERSION};
 use bitcoin::secp256k1::Secp256k1;