+++ src/wallet.rs
@@ -1,4 +1,5 @@
-use crate::errors::Result;
+use anyhow::Result;
+//use crate::errors::Result;
 use structopt::StructOpt;
 use terra_rust_api::Terra;
 use terra_rust_wallet::Wallet;