+++ terra-rust-api/src/client/auth.rs
@@ -1,5 +1,5 @@
 use crate::client::auth_types::AuthAccountResult;
-use crate::errors::Result;
+//use crate::errors::Result;
 use crate::Terra;
 
 pub struct Auth<'a> {
@@ -9,7 +9,7 @@ impl Auth<'_> {
     pub fn create<'a>(terra: &'a Terra) -> Auth<'a> {
         Auth { terra }
     }
-    pub async fn account(&self, account_address: &str) -> Result<AuthAccountResult> {
+    pub async fn account(&self, account_address: &str) -> anyhow::Result<AuthAccountResult> {
         let response = self
             .terra
             .send_cmd::<AuthAccountResult>(&format!("/auth/accounts/{}", account_address), None)