+++ terra-rust-api/src/client/market.rs
@@ -1,7 +1,6 @@
 use crate::client::core_types::Coin;
 use crate::client::market_types::SwapResult;
 
-use crate::errors::Result;
 use crate::Terra;
 
 pub struct Market<'a> {
@@ -11,7 +10,7 @@ impl Market<'_> {
     pub fn create<'a>(terra: &'a Terra) -> Market<'a> {
         Market { terra }
     }
-    pub async fn swap(&self, offer: &Coin, ask_denom: &str) -> Result<SwapResult> {
+    pub async fn swap(&self, offer: &Coin, ask_denom: &str) -> anyhow::Result<SwapResult> {
         let response = self
             .terra
             .send_cmd::<SwapResult>(