+++ terra-rust-api/src/client/tendermint.rs
@@ -1,5 +1,4 @@
 use crate::client::tendermint_types::BlockResult;
-use crate::errors::Result;
 use crate::Terra;
 
 pub struct Tendermint<'a> {
@@ -9,14 +8,14 @@ impl Tendermint<'_> {
     pub fn create<'a>(terra: &'a Terra) -> Tendermint<'a> {
         Tendermint { terra }
     }
-    pub async fn blocks(&self) -> Result<BlockResult> {
+    pub async fn blocks(&self) -> anyhow::Result<BlockResult> {
         let response = self
             .terra
             .send_cmd::<BlockResult>("/blocks/latest", None)
             .await?;
         Ok(response)
     }
-    pub async fn blocks_at_height(&self, height: u64) -> Result<BlockResult> {
+    pub async fn blocks_at_height(&self, height: u64) -> anyhow::Result<BlockResult> {
         let response = self
             .terra
             .send_cmd::<BlockResult>(&format!("/blocks/{}", height), None)