+++ terra-rust-api/src/errors.rs
@@ -1,13 +1,13 @@
 #![allow(missing_docs)]
-use error_chain::error_chain;
-
+use thiserror::Error;
+/*
 impl From<Box<dyn std::error::Error>> for Error {
     fn from(e: Box<dyn std::error::Error>) -> Self {
         Self::from(format!("{:?}", e))
     }
 }
 
-impl From<Box<dyn std::error::Error + Sync + Send>> for Error {
+ impl From<Box<dyn std::error::Error + Sync + Send>> for Error {
     fn from(e: Box<dyn std::error::Error + Sync + Send>) -> Self {
         Self::from(format!("{:?}", e))
     }
@@ -19,6 +19,9 @@ impl PartialEq for Error {
         false
     }
 }
+
+ */
+/*
 error_chain! {
 
     foreign_links {
@@ -64,4 +67,48 @@ error_chain! {
             description("coin parse error")
         }
     }
+}
+
+ */
+#[derive(Error, Debug)]
+pub enum TerraRustAPIError {
+    #[error("Reqwest HTTP(s) Error")]
+    ReqwestError(#[from] ::reqwest::Error),
+    #[error("Terra `{0}` CLI Error")]
+    Terra(String),
+    #[error("Bech32 Decode Error")]
+    Bech32DecodeErr,
+    #[error("Bech32 Decode Error: Key Failed prefix {0} or length {1} Wanted:{2}/{3}")]
+    Bech32DecodeExpanded(String, usize, String, usize),
+    #[error("Mnemonic - Bad Phrase")]
+    Phrasing,
+    #[error("Mnemonic - Missing Phrase")]
+    MissingPhrase,
+    #[error("Bad Implementation. Missing Component")]
+    Implementation,
+    #[error("Unable to convert into public key `{key}`")]
+    Conversion {
+        key: String,
+        source: bitcoin::bech32::Error,
+    },
+    #[error("83 length-missing SECP256K1 prefix")]
+    ConversionSECP256k1,
+    #[error("82 length-missing ED25519 prefix")]
+    ConversionED25519,
+    #[error("Expected Key length of 82 or 83 length was {0}")]
+    ConversionLength(usize),
+    #[error("Expected ED25519 key of length 32 with a BECH32 ED25519 prefix of 5 chars - Len {0} - Hex {1}")]
+    ConversionPrefixED25519(usize, String),
+    #[error("Can't call Transactions without some gas rules")]
+    NoGasOpts,
+    #[error("Can't parse `{parse}` into a coin")]
+    CoinParseErrV {
+        parse: String,
+        source: anyhow::Error,
+    },
+    #[error("Can't parse `{0}` into a coin")]
+    CoinParseErr(String),
+
+    #[error("unknown Terra-Rust API error")]
+    Unknown,
 }